# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderItemRequest:
    properties:
      combo_id:
        type: string
      quantity:
        maximum: 20
        minimum: 1
        type: integer
    required:
    - combo_id
    - quantity
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderItemResponse:
    properties:
      combo_id:
        type: string
      combo_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
      unit_price:
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderResponse:
    properties:
      booking_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderItemResponse'
        type: array
      status:
        type: string
      total:
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboResponse:
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateComboOrderRequest:
    properties:
      booking_id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderItemRequest'
        minItems: 1
        type: array
    required:
    - items
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateComboRequest:
    properties:
      active:
        description: |-
          Active defaults to TRUE when omitted: a product created from the admin
          screen is normally meant to go on sale immediately.
        type: boolean
      description:
        maxLength: 2000
        type: string
      image_url:
        maxLength: 512
        type: string
      name:
        example: Combo 1 - Bap ngot + 2 Pepsi
        maxLength: 255
        minLength: 2
        type: string
      price:
        example: 89000
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateComboRequest:
    properties:
      active:
        type: boolean
      description:
        maxLength: 2000
        type: string
      image_url:
        maxLength: 512
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      price:
        minimum: 0
        type: integer
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: success
        type: string
      reason:
        description: |-
          Reason is a stable machine-readable key for an error ("discount_expired"), so the
          frontend can translate without matching on the English sentence. Absent on success.
        type: string
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 42
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta'
    type: object
...
info:
  contact: {}
  description: API for the Cinema Project box-office system.
  title: BackEnd-CP API
  version: "1.0"
paths:
...
  /admin/concessions:
    get:
      description: Unlike the public GET /combos this is paged and INCLUDES inactive
        products, so an operator can put one back on sale.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Match name or description
        in: query
        name: search
        type: string
      - description: Filter on sale / off sale; omit for both
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List the concession catalogue (operator)
      tags:
      - combos
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateComboRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Add a concession product
      tags:
      - combos
  /admin/concessions/{id}:
    delete:
      description: 'Soft delete: past combo orders still reference it, so the row
        stays and only leaves the catalogue.'
      parameters:
      - description: Combo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Retire a concession product
      tags:
      - combos
    get:
      parameters:
      - description: Combo id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get one concession product (operator)
      tags:
      - combos
    patch:
      consumes:
      - application/json
      description: 'PARTIAL update: an omitted field is left alone. A body with no
        field at all is 400/40001.'
      parameters:
      - description: Combo id
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateComboRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Update a concession product
      tags:
      - combos
...
  /combo-orders:
    post:
      consumes:
      - application/json
      description: Independent of the ticket booking flow; booking_id is an optional
        correlation only.
      parameters:
      - description: Combo order
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CreateComboOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: booking_id does not belong to the caller
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Place a combo/concession order
      tags:
      - combos
  /combo-orders/me:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.ComboOrderResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List the current user's combo orders
      tags:
      - combos
...
securityDefinitions:
  BearerAuth:
    description: 'Header format: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"